+++ temas/abstraccion/csharp/Code/Libro.cs
