+++ temas/abstraccion/csharp/Code/Revista.cs
